Fixed texture uint32 overflow for upload buffer
Need to use multiple buffers to make it work

+++ src/Runtime/D3D12/D3D/Material.h
@@ -6,7 +6,7 @@
 class Material
 {
 	public:
-		inline uint32_t TextureSize()
+		inline UINT64 TextureSize()
 		{
 			return m_Textures.GetTotalTextureSize();
 		}
